% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{compute_expected_survival}
\alias{compute_expected_survival}
\title{Compute Expected Survival Probabilities}
\usage{
compute_expected_survival(data, cox_model, fixed_time)
}
\description{
Compute Expected Survival Probabilities
}
\keyword{internal}
