% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Survival Metrics with Optional Bayesian Adjustment}
\usage{
compute_metrics(
  center_summary,
  bayesian_adjust,
  bayesian_method,
  bayesian_model = NULL
)
}
\description{
Compute Survival Metrics with Optional Bayesian Adjustment
}
\keyword{internal}
