% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{compute_observed_survival}
\alias{compute_observed_survival}
\title{Compute Observed Survival Probabilities}
\usage{
compute_observed_survival(data, fixed_time)
}
\description{
Compute Observed Survival Probabilities
}
\keyword{internal}
