% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{fit_bayesian_hierarchical}
\alias{fit_bayesian_hierarchical}
\title{Fit Bayesian Hierarchical Model}
\usage{
fit_bayesian_hierarchical(data)
}
\description{
Fit Bayesian Hierarchical Model
}
\keyword{internal}
