% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{fit_cox_model}
\alias{fit_cox_model}
\title{Fit Cox Proportional Hazards Model}
\usage{
fit_cox_model(data, numeric_covariates, factor_covariates)
}
\description{
Fit Cox Proportional Hazards Model
}
\keyword{internal}
