% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{generate_funnel_limits}
\alias{generate_funnel_limits}
\title{Generate Funnel Plot Limits}
\usage{
generate_funnel_limits(metrics, conf_levels)
}
\description{
Generate Funnel Plot Limits
}
\keyword{internal}
