% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Survival Analysis}
\usage{
prepare_data(
  data,
  center_col,
  status_col,
  time_col,
  numeric_covariates,
  factor_covariates
)
}
\description{
Prepare Data for Survival Analysis
}
\keyword{internal}
