% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{summarize_centers}
\alias{summarize_centers}
\title{Summarize Center-Level Survival Metrics}
\usage{
summarize_centers(data, expected_survival, observed_survival)
}
\description{
Summarize Center-Level Survival Metrics
}
\keyword{internal}
