% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Survival Metrics with Optional Empirical Bayes Adjustment}
\usage{
compute_metrics(center_summary, bayesian_adjust)
}
\description{
Compute Survival Metrics with Optional Empirical Bayes Adjustment
}
\keyword{internal}
