% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_survival.R
\name{create_funnel_plot}
\alias{create_funnel_plot}
\title{Create Funnel Plot}
\usage{
create_funnel_plot(
  metrics,
  limits,
  highlight = NULL,
  conf_levels,
  include_legend = TRUE
)
}
\description{
Create Funnel Plot
}
\keyword{internal}
